requirements.txt
WebOb
Paste
webapp2
GoogleAppEngineCloudStorageClient
